requirements.txt
Django==5.2.12
django-allauth==65.15.0
django-environ==0.13.0
python-dotenv==1.2.2
requests==2.32.5
PyJWT==2.12.1
cryptography==46.0.5
WeasyPrint==68.1
whitenoise==6.12.0
Pillow==12.1.1
waitress==3.0.2
xhtml2pdf==0.2.17
pytest==9.0.2
pytest-django==4.12.0
